requirements.txt
selenium==4.11.2
webdriver-manager==4.2.3
requests==2.31.0
gspread==5.10.0
google-auth==2.24.0
python-dotenv==1.0.1
beautifulsoup4==4.12.2
pandas==2.2.3
